Refactor the homepage code out into its own module for cleanliness
Excluding the homepage from the coverage since it's covered by acceptance tests
and will ultimately always be bespoke queries and data munging necessary to
generate a pretty set of data for passers-by
